+++ app/models/claim_decision.rb
@@ -7,6 +7,8 @@ class ClaimDecision < ApplicationRecord
     true
   end
 
+  scope :yesterday, -> { where(decision_date: 1.day.ago.all_day) }
+
   def self.to_csv
     CSV.generate(headers: true) do |csv|
       csv << attribute_names