+++ app/models/early_career_payments.rb
@@ -40,6 +40,10 @@ module EarlyCareerPayments
     "https://docs.google.com/forms/TO-BE-REPLACED-by-response-to-ECP-509/viewform"
   end
 
+  def feedback_email
+    "earlycareerteacherpayments@digital.education.gov.uk"
+  end
+
   def short_name
     I18n.t("early_career_payments.policy_short_name")
   end