+++ spec/features/ineligible_early_career_payments_claims_spec.rb
@@ -3,7 +3,7 @@ require "rails_helper"
 RSpec.feature "Ineligible Teacher Early-Career Payments claims" do
   scenario "NQT not in Academic Year after ITT" do
     visit landing_page_path(EarlyCareerPayments.routing_name)
-    expect(page).to have_link(href: EarlyCareerPayments.feedback_url)
+    expect(page).to have_link(href: "mailto:#{EarlyCareerPayments.feedback_email}")
 
     # - Landing (start)
     expect(page).to have_text(I18n.t("early_career_payments.landing_page"))