+++ spec/models/early_career_payments/admin_tasks_presenter_spec.rb
@@ -12,10 +12,13 @@ RSpec.describe EarlyCareerPayments::AdminTasksPresenter, type: :model do
         :early_career_payments_eligibility,
         current_school: school,
         eligible_itt_subject: :mathematics,
-        itt_academic_year: AcademicYear::Type.new.serialize(AcademicYear.new(2018))
+        itt_academic_year: AcademicYear::Type.new.serialize(AcademicYear.new(2018)),
+        qualification: qualification
       )
     )
   end
+
+  let(:qualification) { :undergraduate_itt }
 
   subject(:presenter) { described_class.new(claim) }
 
@@ -39,7 +42,8 @@ RSpec.describe EarlyCareerPayments::AdminTasksPresenter, type: :model do
   describe "#qualifications" do
     it "returns an array of label and values for displaying information for qualification checks" do
       expected_array = [
-        ["ITT start/end year", "In the academic year 2018 to 2019"],
+        ["Qualification", "Undergraduate ITT"],
+        ["ITT end year", "In the academic year 2018 to 2019"],
         ["ITT subject", "Mathematics"]
       ]
 
@@ -47,8 +51,31 @@ RSpec.describe EarlyCareerPayments::AdminTasksPresenter, type: :model do
     end
 
     it "sets the “Award year” value based on the academic year the claim was made in" do
-      expected_qts_answer = presenter.qualifications[0][1]
+      expected_qts_answer = presenter.qualifications[1][1]
       expect(expected_qts_answer).to eq "In the academic year 2018 to 2019"
+    end
+
+    [
+      {qualification: :assessment_only, qualification_text: "Assessment only", year_text: "end"},
+      {qualification: :overseas_recognition, qualification_text: "Overseas recognition", year_text: "end"},
+      {qualification: :postgraduate_itt, qualification_text: "Postgraduate ITT", year_text: "start"},
+      {qualification: :undergraduate_itt, qualification_text: "Undergraduate ITT", year_text: "end"}
+    ].each do |spec|
+      context "with qualification #{spec[:qualification]}" do
+        let(:qualification) { spec[:qualification] }
+
+        it "returns array with qualification #{spec[:qualification_text]}" do
+          expect(presenter.qualifications).to include(
+            ["Qualification", spec[:qualification_text]]
+          )
+        end
+
+        it "returns array with year #{spec[:year_text]}" do
+          expect(presenter.qualifications).to include(
+            ["ITT #{spec[:year_text]} year", "In the academic year 2018 to 2019"]
+          )
+        end
+      end
     end
   end
 end